Moved ProviderSelectionController
