--- app/controllers/candidate_interface/continuous_applications/course_choices/provider_selection_controller.rb
@@ -1,23 +1,0 @@
-module CandidateInterface
-  module ContinuousApplications
-    module CourseChoices
-      class ProviderSelectionController < CandidateInterface::CourseChoices::BaseController
-      private
-
-        def step_params
-          params[:provider_id].present? ? provider_params : params
-        end
-
-        def provider_params
-          ActionController::Parameters.new({
-            current_step => { provider_id: params[:provider_id] },
-          })
-        end
-
-        def current_step
-          :provider_selection
-        end
-      end
-    end
-  end
-end
