+++ config/routes/candidate.rb
@@ -341,8 +341,8 @@ namespace :candidate_interface, path: '/candidate' do
 
       get '/go-to-find' => 'course_choices/go_to_find#new', as: :continuous_applications_go_to_find_explanation
 
-      get '/provider' => 'continuous_applications/course_choices/provider_selection#new', as: :continuous_applications_provider_selection
-      post '/provider' => 'continuous_applications/course_choices/provider_selection#create'
+      get '/provider' => 'course_choices/provider_selection#new', as: :continuous_applications_provider_selection
+      post '/provider' => 'course_choices/provider_selection#create'
 
       get '/provider/:provider_id/course' => 'continuous_applications/course_choices/which_course_are_you_applying_to#new', as: :continuous_applications_which_course_are_you_applying_to
       post '/provider/:provider_id/course' => 'continuous_applications/course_choices/which_course_are_you_applying_to#create'
